refactor: domain/models/ 하위 폴더로 models 분리
