+++ docs/REQUIREMENTS.md
@@ -34,6 +34,8 @@
 - Rank 객체로부터 총 상금을 구한다.
 - 총 구매액을 구한다.
 
+[V] domain > models 폴더로 모델들 분리하기
+
 (UI)
 
 - 총 수익률은 소수 둘째 자리 반올림하고 %를 붙인다.
