+++ src/js/domain/models/Lotto/errors.js
