+++ src/js/domain/models/Lotto/index.js
