+++ src/js/domain/models/LottoMachine.js
