+++ src/js/domain/models/Rank/errors.js
