+++ src/js/domain/models/Rank/index.js
