+++ src/js/domain/models/WinningLotto/errors.js
