+++ src/js/domain/models/WinningLotto/index.js
